+++ aiotdlib/api/errors/__init__.py
@@ -1,1 +1,5 @@
-from .error import AioTDLibError, BadRequest, Unauthorized
+from .error import (
+    AioTDLibError,
+    BadRequest,
+    Unauthorized,
+)