+++ generator/parser/__init__.py
@@ -1,2 +1,5 @@
+from .entities import (
+    Constructor,
+    Method,
+)
 from .parser import Parser
-from .entities import Constructor, Method